pyproject.toml
[project]
name = "openplan"
version = "0.1.0"
description = "A deterministic planning harness for OpenSpec/OpenCode"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "jinja2",
    "typer",
    "pyyaml",
    "rich",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]

[project.scripts]
openplan = "openplan.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["openplan"]
